Added v2.0 final Backend

+++ project_spring_boot_backend/src/main/java/com/app/entities/SearchRestByCuisineId.java
@@ -17,10 +17,10 @@ import lombok.Setter;
 public class SearchRestByCuisineId implements Serializable{
 	@Column(name="cuisine_id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer cuisineId ;
+	private Long cuisineId ;
 	
 	@Column(name="rest_id")
-	private Integer restId ;
+	private Long restId ;
 	
 	@Override
     public boolean equals(Object o) {
